Refactor test classes to use identity common testutils (#670)
