+++ components/org.wso2.carbon.identity.oauth/pom.xml
@@ -249,12 +249,10 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.wso2.carbon.identity</groupId>
-            <artifactId>org.wso2.carbon.identity.common.testng</artifactId>
-            <version>5.5.115-SNAPSHOT</version>
+            <groupId>org.wso2.carbon.identity.inbound.auth.oauth2</groupId>
+            <artifactId>org.wso2.carbon.identity.oauth.common.testng</artifactId>
             <scope>test</scope>
         </dependency>
-
     </dependencies>
 
     <build>
