+++ components/org.wso2.carbon.identity.webfinger/pom.xml
@@ -81,11 +81,6 @@
             <artifactId>powermock-api-mockito</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity</groupId>
-            <artifactId>org.wso2.carbon.identity.common.testng</artifactId>
-            <version>5.5.115-SNAPSHOT</version>
-        </dependency>
     </dependencies>
     <build>
         <plugins>
