+++ components/org.wso2.carbon.identity.webfinger/src/test/java/org/wso2/carbon/identity/webfinger/DefaultWebFingerProcessorTest.java
@@ -1,3 +1,21 @@
+/*
+ * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package org.wso2.carbon.identity.webfinger;
 
 import org.apache.commons.collections.iterators.IteratorEnumeration;
@@ -7,9 +25,8 @@ import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import org.wso2.carbon.identity.test.common.testng.WithAxisConfiguration;
-import org.wso2.carbon.identity.test.common.testng.WithCarbonHome;
-import org.wso2.carbon.identity.test.common.testng.WithRealmService;
+import org.wso2.carbon.identity.common.testng.WithCarbonHome;
+import org.wso2.carbon.identity.common.testng.WithRealmService;
 import org.wso2.carbon.identity.webfinger.internal.WebFingerServiceComponentHolder;
 
 import java.util.HashMap;
@@ -25,7 +42,6 @@ import static org.testng.Assert.fail;
  * Tests for DefaultWebFingerProcessor.
  */
 @WithCarbonHome
-@WithAxisConfiguration
 @WithRealmService(injectToSingletons = { WebFingerServiceComponentHolder.class })
 public class DefaultWebFingerProcessorTest {
