--- test-utils/org.wso2.carbon.identity.common.testng/src/main/java/org/wso2/carbon/identity/test/common/testng/WithRealmService.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.test.common.testng;
-
-import org.wso2.carbon.utils.multitenancy.MultitenantConstants;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface WithRealmService {
-    int tenantId() default MultitenantConstants.SUPER_TENANT_ID;
-    String tenantDomain() default MultitenantConstants.SUPER_TENANT_DOMAIN_NAME;
-    Class[] injectToSingletons() default {};
-    boolean initUserStoreManager() default false;
-}
