--- test-utils/org.wso2.carbon.identity.common.testng/src/main/java/org/wso2/carbon/identity/test/common/testng/logger/NullAppender.java
@@ -1,25 +1,0 @@
-package org.wso2.carbon.identity.test.common.testng.logger;
-
-import org.apache.log4j.AppenderSkeleton;
-import org.apache.log4j.spi.LoggingEvent;
-
-/**
- * Log4J Log Appender, which does nothing. Logs simply goes into a black hole.
- */
-public class NullAppender extends AppenderSkeleton {
-
-    @Override
-    protected void append(LoggingEvent loggingEvent) {
-
-    }
-
-    @Override
-    public void close() {
-
-    }
-
-    @Override
-    public boolean requiresLayout() {
-        return false;
-    }
-}
