--- test-utils/org.wso2.carbon.identity.common.testng/src/main/resources/META-INF/services/org.testng.ITestNGListener
@@ -1,2 +1,0 @@
-org.wso2.carbon.identity.test.common.testng.LogLevelChangeListener
-org.wso2.carbon.identity.test.common.testng.CarbonBasedTestListener
