+++ test-utils/org.wso2.carbon.identity.oauth.common.testng/src/main/java/org/wso2/carbon/identity/test/common/testng/utils/ReadCertStoreSampleUtil.java
